added pretraining
