+++ multimodal/src/autogluon/multimodal/optimization/utils.py
@@ -1014,158 +1014,3 @@ def compute_probability(
         prob = 1 - prob
 
     return prob
-
-
-class ReconstructionLoss:
-    def __init__(self, model):
-        self.model = model
-
-    def __call__(self, batch, batch_):
-        batch_ = batch_[self.model.prefix]["logits"]
-        loss = 0
-        for permodel in self.model.model:
-            if hasattr(permodel, "categorical_key"):
-                for y_, y in zip(batch_["cat_out"], batch[permodel.categorical_key]):
-                    loss += F.cross_entropy(y_, y.long())
-            if hasattr(permodel, "numerical_key"):
-                y = batch[permodel.numerical_key]
-                y_ = batch_["num_out"]
-                loss += F.mse_loss(y_, y)
-        return loss
-
-
-class ContrastiveTransformations:
-    def __init__(self, model, mode, problem_type, corruption_rate):
-        self.model = model
-        self.mode = mode if mode is not None else "identical"
-        self.problem_type = problem_type
-        self.corruption_rate = corruption_rate
-        self.last_batch = None
-
-    def __call__(self, batch):
-        if self.mode == "identical":
-            return self.identical(batch)
-        elif self.mode == "permutation":
-            return self.random_perm(batch)
-        else:
-            raise ValueError(
-                f"Current mode {self.mode} is not supported."
-                "Consider choosing from the following options:"
-                "identical, permutation."
-            )
-
-    def identical(self, batch):
-        batch = copy.deepcopy(batch)
-        return batch
-
-    def random_perm(self, batch):
-        corruption_rate = self.corruption_rate
-        (batch_size,) = batch[self.model.label_key].size()
-        batch = copy.deepcopy(batch)
-
-        num_features = 0
-        for permodel in self.model.model:
-            if hasattr(permodel, "categorical_key"):
-                num_features += len(batch[permodel.categorical_key])
-            if hasattr(permodel, "numerical_key"):
-                _, m = batch[permodel.numerical_key].size()
-                num_features += m
-
-        corruption_mask = torch.zeros(
-            batch_size, num_features, dtype=torch.bool, device=batch[self.model.label_key].device
-        )
-        corruption_len = int(num_features * corruption_rate)
-        for i in range(batch_size):
-            corruption_idx = torch.randperm(num_features)[:corruption_len]
-            corruption_mask[i, corruption_idx] = True
-        feature_idx = 0
-
-        for permodel in self.model.model:
-            if hasattr(permodel, "categorical_key"):
-                categorical_features = []
-                for categorical_feature in batch[permodel.categorical_key]:
-                    random_idx = torch.randint(high=batch_size, size=(batch_size,))
-                    random_sample = categorical_feature[random_idx].clone()
-                    positive = torch.where(corruption_mask[:, feature_idx], random_sample, categorical_feature)
-                    feature_idx += 1
-                    categorical_features.append(positive)
-                batch[permodel.categorical_key] = tuple(categorical_features)
-            if hasattr(permodel, "numerical_key"):
-                numerical_features = batch[permodel.numerical_key]
-                _, m = numerical_features.size()
-                indices = torch.randint(high=batch_size, size=(batch_size, m))
-                random_sample = numerical_features[indices, torch.arange(m).unsqueeze(0)].clone()
-                batch[permodel.numerical_key] = torch.where(
-                    corruption_mask[:, feature_idx : feature_idx + m], random_sample, numerical_features
-                )
-                feature_idx += m
-        return batch
-
-
-class NTXent(nn.Module):
-    def __init__(self, temperature=0.1):
-        """NT-Xent loss for contrastive learning using cosine distance as similarity metric as used in [SimCLR](https://arxiv.org/abs/2002.05709).
-        Implementation adapted from https://theaisummer.com/simclr/#simclr-loss-implementation
-        Args:
-            temperature (float, optional): scaling factor of the similarity metric. Defaults to 1.0.
-        """
-        super().__init__()
-        self.temperature = temperature
-
-    def forward(self, z_i, z_j):
-        """Compute NT-Xent loss using only anchor and positive batches of samples. Negative samples are the 2*(N-1) samples in the batch
-        Args:
-            z_i (torch.tensor): anchor batch of samples
-            z_j (torch.tensor): positive batch of samples
-        Returns:
-            float: loss
-        """
-        batch_size = z_i.size(0)
-        z_i, z_j = z_i.flatten(1), z_j.flatten(1)
-
-        # compute similarity between the sample's embedding and its corrupted view
-        z = torch.cat([z_i, z_j], dim=0)
-        similarity = F.cosine_similarity(z.unsqueeze(1), z.unsqueeze(0), dim=2)
-
-        sim_ij = torch.diag(similarity, batch_size)
-        sim_ji = torch.diag(similarity, -batch_size)
-        positives = torch.cat([sim_ij, sim_ji], dim=0)
-
-        # mask = (~torch.eye(batch_size * 2, batch_size * 2, dtype=torch.bool)).float().to(z_i.get_device())
-        numerator = torch.exp(positives / self.temperature)
-
-        denominator = torch.exp(similarity / self.temperature)  # * mask
-        all_losses = -torch.log(numerator / torch.mean(denominator, dim=1))
-
-        # all_losses = -torch.log(numerator)
-        loss = torch.sum(all_losses) / (2 * batch_size)
-
-        return loss
-
-
-class DistillLoss(nn.Module):
-    def __init__(self, temperature=1):
-        """NT-Xent loss for contrastive learning using cosine distance as similarity metric as used in [SimCLR](https://arxiv.org/abs/2002.05709).
-        Implementation adapted from https://theaisummer.com/simclr/#simclr-loss-implementation
-        Args:
-            temperature (float, optional): scaling factor of the similarity metric. Defaults to 1.0.
-        """
-        super().__init__()
-        self.temperature = temperature
-
-    def forward(self, z_i, z_j):
-        """Compute NT-Xent loss using only anchor and positive batches of samples. Negative samples are the 2*(N-1) samples in the batch
-        Args:
-            z_i (torch.tensor): anchor batch of samples
-            z_j (torch.tensor): positive batch of samples
-        Returns:
-            float: loss
-        """
-        z_i, z_j = z_i.flatten(1), z_j.flatten(1)
-
-        if z_i.size(1) == 1:
-            return F.mse_loss(z_i, z_j)
-        else:
-            z_i, z_j = z_i / self.temperature, z_j / self.temperature
-            z_i = F.softmax(z_i, dim=-1)
-            return F.cross_entropy(z_j, z_i)
